requirements.txt
akismet==1.0.1
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.4
django-allauth==0.33.0
django-appconf==1.0.2
django-avatar==4.0.1
django-bootstrap-form==3.3
django-bootstrap3==9.0.0
django-braces==1.11.0
django-debug-toolbar==1.8
django-envelope==1.3
django-extensions==1.9.0
django-friendship==1.5.0
django-model-utils==3.0.0
django-search-views==0.3.7
django-star-ratings==0.5.6
django-stronghold==0.2.9
djangorestframework==3.6.4
gunicorn==19.7.1
idna==2.6
nltk==3.2.4
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.2.3
urllib3==1.22
whitenoise==3.3.0
